requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
charset-normalizer==2.1.0
cycler==0.11.0
einops==0.3.2
fairscale==0.4.6
fonttools==4.34.2
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
google-auth==2.9.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
imageio==2.10.4
imageio-ffmpeg==0.4.5
immutabledict==2.2.1
importlib-metadata==4.12.0
kiwisolver==1.4.3
kornia==0.6.1
Markdown==3.3.7
matplotlib==3.5.2
multidict==6.0.2
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.5.4.58
packaging==21.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.6.4
pytorch-ranger==0.1.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.10.0
torch-optimizer==0.3.0
torchmetrics==0.9.2
torchsummary==1.5.1
torchvision==0.11.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
